admin api list all customers done

+++ src/main/java/com/ttn/bootcampProject/repos/SellerRepository.java
@@ -3,5 +3,7 @@ package com.ttn.bootcampProject.repos;
 import com.ttn.bootcampProject.entities.Seller;
 import org.springframework.data.repository.CrudRepository;
 
-public interface SellerRepository extends CrudRepository<Seller, Integer> {
+public interface SellerRepository extends CrudRepository<Seller, Long> {
+
+
 }
